tests/wpt/web-platform-tests/service-workers/service-worker/fetch-request-css-images.https.html

<!DOCTYPE html>
<title>Service Worker: FetchEvent for css image</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js?pipe=sub"></script>
<script>
var url_count = 0;
var expected_results = {};

function css_image_test(frame, url, type, expexted_mode,
                        expected_credentials) {
  var actual_url = url + (++url_count);
  expected_results[actual_url] = {
      url: actual_url,
      mode: expexted_mode,
      credentials: expected_credentials,
      message: 'CSSImage load (url:' + actual_url + ' type:' + type + ')'
    };
  return frame.contentWindow.load_css_image(actual_url, type);
}

function css_image_set_test(frame, url, type, expexted_mode,
                            expected_credentials) {
  var actual_url = url + (++url_count);
  expected_results[actual_url] = {
      url: actual_url,
      mode: expexted_mode,
      credentials: expected_credentials,
      message: 'CSSImageSet load (url:' + actual_url + ' type:' + type + ')'
    };
  return frame.contentWindow.load_css_image_set(actual_url, type);
}

async_test(function(t) {
    var SCOPE = 'resources/fetch-request-resources-iframe.https.html';
    var SCRIPT = 'resources/fetch-request-resources-worker.js';
    var host_info = get_host_info();
    var LOCAL_URL =
      host_info['HTTPS_ORIGIN'] + base_path() + 'resources/dummy?test';
    var REMOTE_URL =
      host_info['HTTPS_REMOTE_ORIGIN'] + base_path() + 'resources/dummy?test';
    var worker;
    var frame;
    service_worker_unregister_and_register(t, SCRIPT, SCOPE)
      .then(function(registration) {
          worker = registration.installing;
          return wait_for_state(t, worker, 'activated');
        })
      .then(function() {
          return new Promise(function(resolve) {
              var channel = new MessageChannel();
              channel.port1.onmessage = t.step_func(function(msg) {
                if (msg.data.ready) {
                  resolve();
                  return;
                }
                var result = msg.data;
                var expected = expected_results[result.url];
                if (!expected) {
                  return;
                }
                assert_equals(
                    result.mode, expected.mode,
                    'mode of ' + expected.message +  ' must be ' +
                    expected.mode + '.');
                assert_equals(
                    result.credentials, expected.credentials,
                    'credentials of ' + expected.message +  ' must be ' +
                    expected.credentials + '.');
                --url_count;
                delete expected_results[result.url];
                if (url_count == 0) {
                  frame.remove();
                  service_worker_unregister_and_done(t, SCOPE);
                }
              });
              worker.postMessage(
                {port: channel.port2}, [channel.port2]);
            });
        })
      .then(function() { return with_iframe(SCOPE); })
      .then(function(f) {
        frame = f;

        css_image_test(f, LOCAL_URL, 'backgroundImage', 'no-cors', 'include');
        css_image_test(f, REMOTE_URL, 'backgroundImage', 'no-cors', 'include');

        css_image_test(f, LOCAL_URL, 'shapeOutside', 'cors', 'same-origin');
        css_image_test(f, REMOTE_URL, 'shapeOutside', 'cors', 'same-origin');

        css_image_set_test(f, LOCAL_URL, 'backgroundImage', 'no-cors', 'include');
        css_image_set_test(f, REMOTE_URL, 'backgroundImage', 'no-cors', 'include');
        css_image_set_test(f, LOCAL_URL, 'shapeOutside', 'cors', 'same-origin');
        css_image_set_test(f, REMOTE_URL, 'shapeOutside', 'cors', 'same-origin');
      })
      .catch(unreached_rejection(t));
  }, 'Verify FetchEvent for css images.');
</script>
